Chest spawn actual items
When player opens a chest either a health, ammo, or bomb object will
spawn in front of the chest. Pressing E on either the health or ammo
object will use the item. When a bomb spawns, bomb will explode and
instantly kill the player if it is within the radius.

Added a max ammo 30.
Ammo item gives +5 ammo.
Health item gives +1 health.
Bomb damage is -3 health (instant kill).